Httpimg: rename package.
