+++ contrib/httpimg/httpimg_test.go
@@ -1,9 +1,9 @@
-package fhttp_test
+package httpimg_test
 
 import (
 	"fmt"
 	"github.com/jung-kurt/gofpdf"
-	"github.com/jung-kurt/gofpdf/contrib/fhttp"
+	"github.com/jung-kurt/gofpdf/contrib/httpimg"
 	"path/filepath"
 )
 
@@ -32,7 +32,7 @@ func ExampleFpdf_AddHttpImage() {
 	pdf.AddPage()
 
 	url := "https://github.com/jung-kurt/gofpdf/raw/master/image/logo_gofpdf.jpg?raw=true"
-	fhttp.RegisterHttpImage(pdf, url, "")
+	httpimg.RegisterHttpImage(pdf, url, "")
 	pdf.Image(url, 100, 100, 20, 20, false, "", 0, "")
 
 	fileStr := exampleFilename("Fpdf_AddHttpImage")
