+++ contrib/httpimg/pdf/.gitignore
